requirements.txt
flask==3.0.0
python-pptx==0.6.23
pillow==10.1.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0 